+++ streamparse/ext/invoke.py
@@ -14,6 +14,7 @@ from __future__ import absolute_import, print_function, unicode_literals
 import os
 import re
 import shutil
+import sys
 import time
 from io import open
 from tempfile import NamedTemporaryFile
@@ -33,6 +34,23 @@ __all__ = ["list_topologies", "kill_topology", "run_local_topology",
 # TODO: remove boilerplate get_env_config, get_nimbus_for_env_config...
 # from all these with something like
 # @task("setup")
+
+
+
+def get_user_tasks():
+    """Get tasks defined in a user's tasks.py and fabric.py file which is
+    assumed to be in the current working directory.
+
+    :returns: tuple invoke_tasks, fabric_tasks
+    """
+    try:
+        sys.path.insert(0, os.getcwd())
+        import tasks as user_invoke
+        import fabfile as user_fabric
+        return user_invoke, user_fabric
+    except ImportError:
+        return None, None
+
 
 def is_safe_to_submit(topology_name):
     """Check to see if a topology is currently running or is in the process of
@@ -65,6 +83,7 @@ def _list_topologies(run_args=None, run_kwargs=None):
         run_args = []
     if run_kwargs is None:
         run_kwargs = {}
+    run_kwargs['pty'] = True
     cmd = ["lein",
            "run -m streamparse.commands.list/-main"]
     return run(" ".join(cmd), *run_args, **run_kwargs)
@@ -84,6 +103,7 @@ def _kill_topology(topology_name, run_args=None, run_kwargs=None):
         run_args = []
     if run_kwargs is None:
         run_kwargs = {}
+    run_kwargs['pty'] = True
     cmd = ["lein",
            "run -m streamparse.commands.kill_topology/-main",
            topology_name]
@@ -131,8 +151,19 @@ def run_local_topology(name=None, time=5, par=2, options=None, debug=False):
     cmd.append("-t {}".format(time))
     if debug:
         cmd.append("--debug")
-    cmd.append('--option "topology.workers={}"'.format(par))
-    cmd.append('--option "topology.acker.executors={}"'.format(par))
+    cmd.append("--option 'topology.workers={}'".format(par))
+    cmd.append("--option 'topology.acker.executors={}'".format(par))
+
+    # Python logging settings
+    if not os.path.isdir("logs"):
+        os.makedirs("logs")
+    log_path = os.path.join(os.getcwd(), "logs")
+    print("Routing Python logging to {}.".format(log_path))
+    cmd.append("--option 'streamparse.log.path=\"{}\"'"
+                   .format(log_path))
+    cmd.append("--option 'streamparse.log.level=\"debug\"'")
+
+
     if options is None:
         options = []
     for option in options:
@@ -154,26 +185,24 @@ def submit_topology(name=None, env_name="prod", par=2, options=None,
     env_name, env_config = get_env_config(env_name)
     host, port = get_nimbus_for_env_config(env_config)
 
-    config["virtualenv_specs"] = config["virtualenv_specs"].rstrip("/")
     activate_env(env_name)
-    create_or_update_virtualenvs(name,
-                                 "{}/{}.txt".format(config["virtualenv_specs"],
-                                                    name))
 
-    # TODO: Super hacky way to replace "python" with our venv executable, need
-    # to fix this
-    with open(topology_file, "r") as fp:
-        contents = fp.read()
-    contents = re.sub(r'"python"',
-                     '"{}/{}/bin/python"'
-                      .format(env_config["virtualenv_root"], name),
-                      contents)
-    tmpfile = NamedTemporaryFile(dir=config["topology_specs"])
-    tmpfile.write(contents)
-    tmpfile.flush()
-    print("Created modified topology definition file {}.".format(tmpfile.name))
+    # pre-submit hooks for invoke and fabric
+    user_invoke, user_fabric = get_user_tasks()
+    pre_submit_invoke = getattr(user_invoke, "pre_submit", None)
+    if callable(pre_submit_invoke):
+        pre_submit_invoke(name, env_name, env_config)
+    pre_submit_fabric = getattr(user_fabric, "pre_submit", None)
+    if callable(pre_submit_fabric):
+        pre_submit_fabric(name, env_name, env_config)
 
-    # replaced with /path/to/venv/bin/python instead
+    config["virtualenv_specs"] = config["virtualenv_specs"].rstrip("/")
+
+    create_or_update_virtualenvs(
+        name, "{}/{}.txt".format(config["virtualenv_specs"], name)
+    )
+    python_path = '/'.join([env_config["virtualenv_root"],
+                           name, "bin", "python"])
 
     # Prepare a JAR that doesn't have Storm dependencies packaged
     topology_jar = jar_for_deploy()
@@ -200,26 +229,60 @@ def submit_topology(name=None, env_name="prod", par=2, options=None,
         os.environ["JVM_OPTS"] = " ".join(jvm_opts)
         cmd = ["lein",
                "run -m streamparse.commands.submit_topology/-main",
-               tmpfile.name]
+               topology_file]
         if debug:
             cmd.append("--debug")
-        cmd.append('--option "topology.workers={}"'.format(par))
-        cmd.append('--option "topology.acker.executors={}"'.format(par))
+        cmd.append("--option 'topology.workers={}'".format(par))
+        cmd.append("--option 'topology.acker.executors={}'".format(par))
+        cmd.append("--option 'topology.python.path=\"{}\"'".format(python_path))
+
+        # Python logging settings
+        log_config = env_config.get("log", {})
+        log_path = log_config.get("path") or env_config.get("log_path")
+        print("Routing Python logging to {}.".format(log_path))
+        if log_path:
+            cmd.append("--option 'streamparse.log.path=\"{}\"'"
+                       .format(log_path))
+        if isinstance(log_config.get("max_bytes"), int):
+            cmd.append("--option 'streamparse.log.max_bytes={}'"
+                       .format(log_config["max_bytes"]))
+        if isinstance(log_config.get("backup_count"), int):
+            cmd.append("--option 'streamparse.log.backup_count={}'"
+                       .format(log_config["backup_count"]))
+        if isinstance(log_config.get("level"), basestring):
+            cmd.append("--option 'streamparse.log.level=\"{}\"'"
+                       .format(log_config["level"].lower()))
+
         if options is None:
             options = []
         for option in options:
-            cmd.append('--option {}'.format(option))
+            # XXX: hacky Parse.ly-related workaround; must fix root
+            # issue with -o options and string values
+            if "deployment_stage" in option:
+                key, val = option.split("=")
+                cmd.append("--option '{}=\"{}\"'".format(key, val))
+            else:
+                cmd.append("--option {}".format(option))
         full_cmd = " ".join(cmd)
         print("Running lein command to submit topology to nimbus:")
         print(full_cmd)
         run(full_cmd)
 
-    tmpfile.close()
+        # post-submit hooks for invoke and fabric
+        post_submit_invoke = getattr(user_invoke, "post_submit", None)
+        if callable(post_submit_invoke):
+            post_submit_invoke(name, env_name, env_config)
+        post_submit_fabric = getattr(user_fabric, "post_submit", None)
+        if callable(post_submit_fabric):
+            post_submit_fabric(name, env_name, env_config)
+
 
 @task
-def tail_topology(env_name="prod", pattern=None):
+def tail_topology(topology_name=None, env_name=None, pattern=None):
+    get_topology_definition(topology_name)
     activate_env(env_name)
-    tail_logs(pattern)
+    tail_logs(topology_name, pattern)
+
 
 @task
 def visualize_topology(name=None, flip=False):
@@ -234,4 +297,3 @@ def visualize_topology(name=None, flip=False):
     print("Running lein command to visualize topology:")
     print(full_cmd)
     run(full_cmd)
-